tweet controller is completed

+++ src/main/java/com/workintech/twitter/service/ReplyServiceImpl.java
@@ -2,9 +2,11 @@ package com.workintech.twitter.service;
 
 import com.workintech.twitter.entity.Reply;
 import com.workintech.twitter.entity.Tweet;
+import com.workintech.twitter.exceptions.TwitterException;
 import com.workintech.twitter.repository.ReplyRepository;
 import com.workintech.twitter.repository.TweetRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -27,21 +29,24 @@ public class ReplyServiceImpl implements ReplyService {
     if (foundReply.isPresent()) {
         return foundReply.get();
     }
-    //TODO
-        return null;
+    throw new TwitterException("Reply with given id is not valid: " + replyId, HttpStatus.NOT_FOUND);
     }
 
     @Override
     public void replyToTweet(int tweetId, Reply reply) {
-        Tweet tweet = tweetRepository.findById(tweetId).orElse(null);
-        if (tweet != null) {
-            reply.setTweet(tweet);
+        Optional<Tweet> tweet = tweetRepository.findById(tweetId);
+        if (tweet.isPresent()) {
+            reply.setTweet(tweet.get());
             replyRepository.save(reply);
         }
     }
 
     @Override
     public void deleteReply(int replyId) {
+    Optional<Reply> reply = replyRepository.findById(replyId);
+    if(reply.isPresent()) {
         replyRepository.deleteById(replyId);
+
+    }
     }
 }
